Updated IPBlacklist URL and Version number
Added a pipe in the URL for the blacklist download because pulledpork.pl was parsing IPBLACKLIST as the file name to download instead of ip-filter. This was causing the script to fail because pulledpork.pl issues a croak command if the file fails to download and autosnort won't complete unless it gets a successful completion response from pulledpork.

Also updated the version number from 0.7.4 to 0.8.0 because the version number in the pulledpork repo was updated in Sep 2020 and pulledpork.pl was issuing a croak command when it saw that the version numbers in the .conf file didn't match the .pl file that was updated, again causing the autosnort script to fail.